refactor: 패키지 구조 개선 - config를 common 외부로 이동 (#19)
- config 패키지를 루트 레벨로 이동
- 설정 클래스 접근성 및 가독성 향상
- Spring Boot 표준 구조에 맞게 재구성
