+++ src/main/java/com/shoppingmall/ecommerceapi/config/jpa/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.shoppingmall.ecommerceapi.common.config.jpa;
+package com.shoppingmall.ecommerceapi.config.jpa;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @Configuration
 @EntityScan(basePackages = "com.shoppingmall.ecommerceapi")
-@EnableJpaRepositories(basePackages = "com.shoppingmall.ecommerceapi" )
+@EnableJpaRepositories(basePackages = "com.shoppingmall.ecommerceapi")
 public class JpaConfig {
+
 }
