+++ src/main/java/com/shoppingmall/ecommerceapi/config/swagger/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.shoppingmall.ecommerceapi.common.config.swagger;
+package com.shoppingmall.ecommerceapi.config.swagger;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.v3.core.jackson.ModelResolver;
@@ -8,8 +8,8 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class SwaggerConfig {
 
-    @Bean
-    public ModelResolver modelResolver(ObjectMapper objectMapper){
-        return new ModelResolver(objectMapper);
-    }
+  @Bean
+  public ModelResolver modelResolver(ObjectMapper objectMapper) {
+    return new ModelResolver(objectMapper);
+  }
 }
